properly handle RawPacket samples that are bigger than the driver's internal buffer
There is really no reason to reject those. The device might send
multiple packets at once, and the driver internal buffer is only
required to be as big as the biggest packet (even though
performance would be better if it is a few times as large)

This removes the need for the mHasPacket flag as well.