reallocate mListener for each new driver object
Listeners are owned by the driver, so by doing so we remove the
weird gymnastic forcing to remove the listener before the subclass(es)
delete their drivers.

The only downside of doing so is that the driver keep its listener
attached. Commonly, a new driver gets allocated in configureHook,
which will work as-is under the new scheme. Reusing the same
driver will also work. The only scheme that will change is if
one unsets a driver and starts using it again. Under this usage,
that I've never seen in the wild, the old driver will still be tied
to the I/O monitoring ports and these ports will keep on seeing the
I/O send and received on them. If such usage do exist, tasks that
do so will now need to call detachDriver() to untie the driver
and the I/O ports.