fix handling of processIO on non-FileDescriptorActivity deployments
With FileDescriptorActivity, there is essentially a loop that goes
through the activity itself until there is no data on the file
descriptor. The activity calls select() which triggers if there
is data, and calls updateHook and then processIO.

This does not work for the two other deployment modes (ports and
other types of activites). The base implementation of updateHook
only calls hasIO once, and loops for packets that are buffered. That's
fine if the driver's internal buffer is much bigger than the data
that may be queued on the I/O, but otherwise we end up with a queue that
never empties.

This commit changes the behavior from an if() to a while()