always go through waitRead in hasIO
waitRead is now not 100% equivalent to calling select()/poll() as
it may filter out some errors (as in e.g. the UDP case).

Call waitRead unconditionally in hasIO to account for this. Remove
the first_time case in hasIO since it has even less effect now
than it had before.